Replace recipe with recipe from sampleRecipesData
