+++ test/sampleRecipesData.js
@@ -1,50 +1,3 @@
-const testIngredients = [{
-    "id": 0,
-    "name": "rice",
-    "estimatedCostInCents": 150
-  },
-  {
-    "id": 1,
-    "name": "egg",
-    "estimatedCostInCents": 10
-  },
-  {
-    "id": 2,
-    "name": "avocado",
-    "estimatedCostInCents": 250
-  },
-  {
-    "id": 3,
-    "name": "tomatillo",
-    "estimatedCostInCents": 50
-  },
-  {
-    "id": 4,
-    "name": "garlic",
-    "estimatedCostInCents": 25
-  },
-  {
-    "id": 5,
-    "name": "jalapeno",
-    "estimatedCostInCents": 10
-  },
-  {
-    "id": 6,
-    "name": "cilantro",
-    "estimatedCostInCents": 50
-  },
-  {
-    "id": 7,
-    "name": "soy sauce",
-    "estimatedCostInCents": 5
-  },
-  {
-    "id": 8,
-    "name": "brown sugar",
-    "estimatedCostInCents": 10
-  }
-]
-
 const testRecipes = [{
     "id": 1,
     "image": "https://soufflebombay.com/wp-content/uploads/2017/01/Fried-Egg-Avocado-Rice-Bowl.jpg",
@@ -220,58 +173,5 @@ const testRecipes = [{
   }
 ]
 
-const testUserData = [{
-    "name": "Saige O'Kon",
-    "id": 1,
-    "pantry": [{
-        "ingredient": 0,
-        "amount": 4
-      },
-      {
-        "ingredient": 1,
-        "amount": 10
-      },
-      {
-        "ingredient": 2,
-        "amount": 5
-      }
-    ]
-  },
-  {
-    "name": "Ephraim Goyette",
-    "id": 2,
-    "pantry": [{
-        "ingredient": 1,
-        "amount": 3
-      },
-      {
-        "ingredient": 2,
-        "amount": 7
-      },
-      {
-        "ingredient": 3,
-        "amount": 8
-      },
-      {
-        "ingredient": 4,
-        "amount": 6
-      },
-      {
-        "ingredient": 6,
-        "amount": 10
-      },
-      {
-        "ingredient": 8,
-        "amount": 13
-      }
-    ]
-  }
 
-]
-
-
-export default {
-  testIngredients,
-  testRecipes,
-  testUserData
-};
+export default testRecipes;
