+++ test/sampleUserData.js
@@ -1,0 +1,50 @@
+const testUserData = [{
+    "name": "Saige O'Kon",
+    "id": 1,
+    "pantry": [{
+        "ingredient": 0,
+        "amount": 4
+      },
+      {
+        "ingredient": 1,
+        "amount": 10
+      },
+      {
+        "ingredient": 2,
+        "amount": 5
+      }
+    ]
+  },
+  {
+    "name": "Ephraim Goyette",
+    "id": 2,
+    "pantry": [{
+        "ingredient": 1,
+        "amount": 3
+      },
+      {
+        "ingredient": 2,
+        "amount": 7
+      },
+      {
+        "ingredient": 3,
+        "amount": 8
+      },
+      {
+        "ingredient": 4,
+        "amount": 6
+      },
+      {
+        "ingredient": 6,
+        "amount": 10
+      },
+      {
+        "ingredient": 8,
+        "amount": 13
+      }
+    ]
+  }
+
+]
+
+export default testUserData;
