+++ test/user-test.js
@@ -2,17 +2,20 @@ import { expect } from 'chai';
 
 import User from '../src/user';
 import data from '../src/data/users-data';
+//import { testIngredients, testRecipes, testUserData} from './test-data'
+import testUserData from './sampleUserData';
+import testRecipes from './sampleRecipesData'
 
 describe('User', function() {
   let user;
-  let userInfo;
   let recipe;
 
   beforeEach(function() {
-    userInfo = data.users[0];
-    user = new User(userInfo)
 
-    recipe = {name: 'Chicken Parm', type: ['italian', 'dinner']};
+    //console.log('the user', testUserData)
+    user = new User(testUserData[0]);
+    recipe = testRecipes[0];
+
   });
 
   it('should be a function', function() {
@@ -28,7 +31,19 @@ describe('User', function() {
   });
 
   it('should initialize with a pantry', function() {
-    expect(user.pantry[0].ingredient).to.eq(11477);
+    expect(user.pantry).to.eql([{
+        "ingredient": 0,
+        "amount": 4
+      },
+      {
+        "ingredient": 1,
+        "amount": 10
+      },
+      {
+        "ingredient": 2,
+        "amount": 5
+      }
+    ]);
   });
 
   it('should initialize with an empty favoriteRecipes array', function() {
@@ -41,21 +56,21 @@ describe('User', function() {
 
   it('should be able to save a recipe to favoriteRecipes', function() {
     user.saveRecipe(recipe);
-    expect(user.favoriteRecipes[0].name).to.equal('Chicken Parm');
+    expect(user.favoriteRecipes[0].name).to.equal('Rice bowl with Fried Egg');
   });
 
   it('should be able to decide to cook a recipe', function() {
     user.decideToCook(recipe);
-    expect(user.recipesToCook[0].name).to.equal('Chicken Parm');
+    expect(user.recipesToCook[0].name).to.equal('Rice bowl with Fried Egg');
   });
 
   it('should be able to filter recipes by type', function() {
     user.saveRecipe(recipe);
-    expect(user.filterRecipes('italian')).to.deep.equal([recipe]);
+    expect(user.filterRecipes('breakfast')).to.deep.equal([recipe]);
   });
 
   it('should be able to search recipes by name', function() {
     user.saveRecipe(recipe);
-    expect(user.searchForRecipe('Chicken Parm')).to.deep.equal([recipe]);
+    expect(user.searchForRecipe('Rice bowl with Fried Egg')).to.deep.equal([recipe]);
   });
 });
